requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
trl>=0.7.0
Pillow>=9.0.0
numpy>=1.21.0
evaluate>=0.4.0
tqdm>=4.64.0
bitsandbytes>=0.41.0
tensorboard>=2.12.0
accelerate>=0.20.0